Move PornDetector to Common package
