--- kifi-backend/modules/rover/app/com/keepit/rover/commanders/RoverCommander.scala
@@ -1,8 +1,0 @@
-package com.keepit.rover.commanders
-
-import com.google.inject.Inject
-
-@Inject()
-class RoverCommander {
-
-}
