+++ kifi-backend/modules/scraper/app/com/keepit/controllers/scraper/PornDetectorController.scala
@@ -1,14 +1,11 @@
 package com.keepit.controllers.scraper
 
-import com.keepit.learning.porndetector.PornDetectorFactory
 import com.keepit.common.controller.ScraperServiceController
 import com.google.inject.Inject
+import com.keepit.rover.sensitivity.{ PornDetectorFactory, PornDetectorUtil, PornWordLikelihood, PornWordLikelihoodStore }
 import play.api.mvc.Action
 import play.api.libs.json._
-import com.keepit.learning.porndetector.PornDetectorUtil
-import com.keepit.learning.porndetector.PornWordLikelihoodStore
 import scala.collection.mutable
-import com.keepit.learning.porndetector.PornWordLikelihood
 
 class PornDetectorController @Inject() (
     factory: PornDetectorFactory,
