+++ kifi-backend/modules/scraper/app/com/keepit/scraper/store/ScraperStore.scala
@@ -1,14 +1,10 @@
 package com.keepit.common.store
 
 import com.google.inject.{ Provides, Singleton }
-import com.amazonaws.services.s3.{ AmazonS3Client, AmazonS3 }
-import com.keepit.common.healthcheck.{ SystemAdminMailSender, AirbrakeNotifier }
-import com.keepit.shoebox.ShoeboxServiceClient
-import com.keepit.common.time.Clock
+import com.amazonaws.services.s3.{ AmazonS3 }
+import com.keepit.common.healthcheck.{ AirbrakeNotifier }
 import play.api.Play._
-import com.keepit.common.aws.AwsModule
 import com.keepit.common.logging.AccessLog
-import com.keepit.learning.porndetector._
 import com.google.inject.Provider
 import com.keepit.scraper.embedly._
 
@@ -23,13 +19,6 @@ case class ScraperProdStoreModule() extends ProdStoreModule {
     val base = current.configuration.getString("cdn.base")
     S3ImageConfig(bucket.get, base.get)
 
-  }
-
-  @Singleton
-  @Provides
-  def bayesPornDetectorStore(amazonS3Client: AmazonS3, accessLog: AccessLog): PornWordLikelihoodStore = {
-    val bucketName = S3Bucket(current.configuration.getString("amazon.s3.bayes.porn.detector.bucket").get)
-    new S3PornWordLikelihoodStore(bucketName, amazonS3Client, accessLog)
   }
 
   @Singleton
@@ -55,16 +44,6 @@ case class ScraperDevStoreModule() extends DevStoreModule(ScraperProdStoreModule
   @Provides
   def s3ImageConfig: S3ImageConfig =
     whenConfigured("cdn.bucket")(prodStoreModule.s3ImageConfig).getOrElse(S3ImageConfig("", "http://dev.ezkeep.com:9000", true))
-
-  @Singleton
-  @Provides
-  def bayesPornDetectorStore(amazonS3ClientProvider: Provider[AmazonS3], accessLog: AccessLog) = {
-    whenConfigured("amazon.s3.bayes.porn.detector.bucket")(
-      prodStoreModule.bayesPornDetectorStore(amazonS3ClientProvider.get, accessLog)
-    ).getOrElse(new InMemoryPornWordLikelihoodStore() {
-        override def syncGet(key: String) = Some(PornWordLikelihood(Map("a" -> 1f)))
-      })
-  }
 
   @Singleton
   @Provides
